Ubuntu instructions and command readability
The end-of-sentence dot gets copied with the rest of the command, so it's better to remove it entirely. The command itself ended on the next line so I separated it entirely. As for Ubuntu - there are Firefox, npm and node.js on it, so we might as well add Linux instructions (can be tested by turning on Linux on Windows PC).